Created ProRequest Image

+++ Muscle/src/main/java/Muscle/common/service/S3Service.java
@@ -14,6 +14,8 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
@@ -30,7 +32,7 @@ public class S3Service {
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
-    // 1. 파일 업로드 메서드
+    // 다중 파일 업로드 메서드
     public List<String> uploadFiles(MultipartFile[] files, String dirName) throws IOException {
         List<String> imageUrls = new ArrayList<>();
 
@@ -47,6 +49,22 @@ public class S3Service {
         }
 
         return imageUrls;
+    }
+
+    // 단일 파일 업로드 메서드
+    public String uploadFile(MultipartFile file, String dirName) throws IOException {
+        String url;
+        try {
+                File uploadFile = convert(file)
+                        .orElseThrow(() -> new IllegalArgumentException("MultipartFile -> File convert failed"));
+
+                url = upload(uploadFile, dirName);
+            } catch (AmazonS3Exception e) {
+                throw new RuntimeException("S3에 파일 업로드 중 오류 발생: " + e.getMessage());
+            }
+
+
+        return url;
     }
 
     private String upload(File uploadFile, String dirName){
@@ -90,20 +108,35 @@ public class S3Service {
     }
 
 
-    public void deleteFile(String url){
+    public void deleteFile(String url) {
         String fileName = extractFileNameFromUrl(url);
-
-        // S3 객체 삭제
-        amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, fileName));
-        System.out.println("파일이 삭제되었습니다: " + fileName);
+        try {
+            // 파일 존재 여부 확인
+            if (amazonS3Client.doesObjectExist(bucket, fileName)) {
+                // 파일 삭제 요청
+                amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, fileName));
+                System.out.println("이미지가 성공적으로 삭제되었습니다.");
+            } else {
+                System.out.println("파일을 찾을 수 없습니다: " + fileName);
+            }
+        } catch (Exception e) {
+            System.err.println("이미지 삭제 중 오류가 발생했습니다: " + e.getMessage());
+            throw e;
+        }
     }
 
-    // URL에서 파일 경로 추출
+    // 수정된 extractFileNameFromUrl 메서드
     private String extractFileNameFromUrl(String fileUrl) {
-        // URL 형식: https://bucket-name.s3.region.amazonaws.com/folder/filename
-        // URL에서 버킷 이름과 경로를 제외한 파일 경로를 추출
-        return fileUrl.substring(fileUrl.lastIndexOf("/") + 1);
+        try {
+            // .com/ 이후의 경로 전체를 추출하고 URL 디코딩하여 한글을 포함한 파일명 처리
+            String fileKey = fileUrl.substring(fileUrl.indexOf(".com/") + 5);
+            return URLDecoder.decode(fileKey, StandardCharsets.UTF_8.toString());
+        } catch (Exception e) {
+            throw new RuntimeException("파일 경로 추출 중 오류가 발생했습니다.", e);
+        }
     }
+
+
 
 
 }
